+++ src/index.ts
@@ -5,6 +5,7 @@ import { PropagatorType } from "./enums/propagator-type.enum";
 import { TelemetryErrorCode } from "./enums/telemetry-error-code.enum";
 import { TelemetrySignal } from "./enums/telemetry-signal.enum";
 import TelemetryConfigError from "./errors/telemetry-config.error";
+import { observeConnectionPool } from "./services/connection-pool.service";
 import { currentTraceId, getTracer, withSpan } from "./services/span.service";
 import Telemetry from "./services/telemetry.service";
 
@@ -13,6 +14,7 @@ export {
   ExporterType,
   getTracer,
   InstrumentationName,
+  observeConnectionPool,
   OtlpProtocol,
   PropagatorType,
   Telemetry,